Implemented refactor
