+++ trustgraph-base/trustgraph/schema/__init__.py
@@ -1,17 +1,10 @@
 
-from . types import *
-from . prompt import *
-from . documents import *
-from . models import *
-from . object import *
-from . topic import *
-from . graph import *
-from . retrieval import *
-from . metadata import *
-from . agent import *
-from . lookup import *
-from . library import *
-from . config import *
-from . flows import *
-from . knowledge import *
+# Import core types and primitives
+from .core import *
+
+# Import knowledge schemas
+from .knowledge import *
+
+# Import service schemas
+from .services import *
