+++ trustgraph-base/trustgraph/schema/core/__init__.py
@@ -1,0 +1,3 @@
+from .primitives import *
+from .metadata import *
+from .topic import *
