+++ trustgraph-base/trustgraph/schema/core/primitives.py
