+++ trustgraph-base/trustgraph/schema/core/topic.py
